requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiotinydb==2.0.0
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
discord.py==2.3.0
distlib==0.3.6
filelock==3.12.2
frozenlist==1.3.3
identify==2.5.24
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.1
parsedatetime==2.6
pathspec==0.11.1
pendulum==2.1.2
platformdirs==3.5.3
pre-commit==3.3.2
python-dateutil==2.8.2
pytzdata==2020.1
PyYAML==6.0
ruff==0.0.272
six==1.16.0
tinydb==4.8.0
tinydb-serialization==2.1.0
tomli==2.0.1
virtualenv==20.23.0
yarl==1.9.2
